+++ src/exercises/lesson-06/src/components/TaskFillerButton.jsx
@@ -1,0 +1,10 @@
+export default function TaskFillerButton({ filter, onFilterChange }) {
+  return (
+    <div>
+      <button onClick={() => onFilterChange('all')}>All</button>
+      <button onClick={() => onFilterChange('completed')}>Completed</button>
+      <button onClick={() => onFilterChange('pending')}>Pending</button>
+      <p>Current filter: {filter}</p>
+    </div>
+  );
+}
