+++ src/exercises/lesson-06/src/components/TaskItem.jsx
@@ -1,0 +1,7 @@
+export default function TaskItem({ task }) {
+  return (
+    <li key={task.id}>
+      {task.title} {task.completed ? '✅' : '⏳'}
+    </li>
+  );
+}
