+++ src/exercises/lesson-06/src/components/UserProfile.jsx
@@ -1,0 +1,3 @@
+export default function UserProfile({ name }) {
+  return <h2>Welcome, {name}</h2>;
+}
