+++ src/exercises/lesson-06/src/hooks/useTasks.js
@@ -1,0 +1,21 @@
+import { useState, useEffect } from 'react';
+export function useTasks() {
+  const [tasks, setTasks] = useState([]);
+  const [loading, setLoading] = useState(true);
+
+  //  #1: Data fetching + state + UI logic all mixed together
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setTasks([
+        { id: 1, title: 'Learn React', completed: true },
+        { id: 2, title: 'Refactor code', completed: false },
+        { id: 3, title: 'Organize files', completed: false },
+      ]);
+      setLoading(false);
+    }, 500);
+
+    return () => clearTimeout(timeout);
+  }, []);
+
+  return { tasks, loading }; //sending UP to studentWork.jsx
+}
