+++ src/exercises/lesson-06/studentWork.jsx
@@ -1,33 +1,16 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
+import UserProfile from './src/components/UserProfile';
+import TaskFillerButton from './src/components/TaskFillerButton';
+import TaskItem from './src/components/TaskItem';
+import filterTasks from './src/utils/filterTasks';
+import useTasks from './src/hooks/useTasks';
 
 export default function StudentWork() {
-  const [tasks, setTasks] = useState([]);
   const [filter, setFilter] = useState('all');
-  const [loading, setLoading] = useState(true);
-
-  //  #1: Data fetching + state + UI logic all mixed together
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      setTasks([
-        { id: 1, title: 'Learn React', completed: true },
-        { id: 2, title: 'Refactor code', completed: false },
-        { id: 3, title: 'Organize files', completed: false },
-      ]);
-      setLoading(false);
-    }, 500);
-
-    return () => clearTimeout(timeout);
-  }, []);
+  const { tasks, loading } = useTasks(); //receiving from useTasks
 
   // #2: Filtering logic inside component
-  let visibleTasks = tasks;
-  if (filter === 'completed') {
-    visibleTasks = tasks.filter((task) => task.completed);
-  }
-  if (filter === 'pending') {
-    visibleTasks = tasks.filter((task) => !task.completed);
-  }
-
+  let visibleTasks = filterTasks(tasks, filter);
   if (loading) {
     return <p>Loading tasks...</p>;
   }
@@ -35,22 +18,15 @@ export default function StudentWork() {
   return (
     <div>
       {/* #3: Hardcoded UI, not reusable */}
-      <h2>Welcome, Student</h2>
+      <UserProfile name={'Student'} />
 
       {/* #4: Repeated button JSX */}
-      <div>
-        <button onClick={() => setFilter('all')}>All</button>
-        <button onClick={() => setFilter('completed')}>Completed</button>
-        <button onClick={() => setFilter('pending')}>Pending</button>
-        <p>Current filter: {filter}</p>
-      </div>
+      <TaskFillerButton filter={filter} onFilterChange={setFilter} />
 
       {/* #5: Inline list rendering */}
       <ul>
         {visibleTasks.map((task) => (
-          <li key={task.id}>
-            {task.title} {task.completed ? '✅' : '⏳'}
-          </li>
+          <TaskItem key={task.id} task={task} /> // ← renders whatever filterTasks returned
         ))}
       </ul>
     </div>
